Rename old RPL implementation as rpl-classic
