+++ os/net/ipv6/multicast/esmrf.c
@@ -49,8 +49,8 @@
 #if UIP_CONF_IPV6_RPL_LITE == 1
 #include "net/rpl-lite/rpl.h"
 #else /* UIP_CONF_IPV6_RPL_LITE == 1 */
-#include "net/rpl/rpl.h"
-#include "net/rpl/rpl-private.h"
+#include "net/rpl-classic/rpl.h"
+#include "net/rpl-classic/rpl-private.h"
 #endif /* UIP_CONF_IPV6_RPL_LITE == 1 */
 #include "net/ip/uip.h"
 #include "net/netstack.h"
