+++ os/net/ipv6/multicast/smrf.c
@@ -50,7 +50,7 @@
 #if UIP_CONF_IPV6_RPL_LITE == 1
 #include "net/rpl-lite/rpl.h"
 #else /* UIP_CONF_IPV6_RPL_LITE == 1 */
-#include "net/rpl/rpl.h"
+#include "net/rpl-classic/rpl.h"
 #endif /* UIP_CONF_IPV6_RPL_LITE == 1 */
 #include "net/netstack.h"
 #include "net/packetbuf.h"
