+++ os/net/mac/tsch/tsch-rpl.h
@@ -37,7 +37,7 @@
 #if UIP_CONF_IPV6_RPL_LITE == 1
 #include "net/rpl-lite/rpl.h"
 #else /* UIP_CONF_IPV6_RPL_LITE == 1 */
-#include "net/rpl/rpl.h"
+#include "net/rpl-classic/rpl.h"
 #endif /* UIP_CONF_IPV6_RPL_LITE == 1 */
 #include "net/mac/tsch/tsch-queue.h"
