+++ os/net/rpl-classic/rpl-dag-root.c
@@ -32,9 +32,9 @@
 #include "contiki.h"
 #include "contiki-net.h"
 
-#include "net/rpl/rpl.h"
-#include "net/rpl/rpl-private.h"
-#include "net/rpl/rpl-dag-root.h"
+#include "net/rpl-classic/rpl.h"
+#include "net/rpl-classic/rpl-private.h"
+#include "net/rpl-classic/rpl-dag-root.h"
 #include "net/ipv6/uip-ds6-route.h"
 
 #include <string.h>
