+++ os/net/rpl-classic/rpl-dag.c
@@ -45,7 +45,7 @@
 
 #include "contiki.h"
 #include "net/link-stats.h"
-#include "net/rpl/rpl-private.h"
+#include "net/rpl-classic/rpl-private.h"
 #include "net/ip/uip.h"
 #include "net/ipv6/uip-nd6.h"
 #include "net/ipv6/uip-ds6-nbr.h"
