+++ os/net/rpl-classic/rpl-mrhof.c
@@ -46,8 +46,8 @@
  * @{
  */
 
-#include "net/rpl/rpl.h"
-#include "net/rpl/rpl-private.h"
+#include "net/rpl-classic/rpl.h"
+#include "net/rpl-classic/rpl-private.h"
 #include "net/nbr-table.h"
 #include "net/link-stats.h"
