+++ os/net/rpl-classic/rpl-ns.c
@@ -37,14 +37,14 @@
  * \author Simon Duquennoy <simon.duquennoy@inria.fr>
  */
 
-#include "net/rpl/rpl-conf.h"
+#include "net/rpl-classic/rpl-conf.h"
 
 #include "net/ip/uip.h"
 #include "net/ip/tcpip.h"
 #include "net/ipv6/uip-ds6.h"
 #include "net/ipv6/uip-icmp6.h"
-#include "net/rpl/rpl-private.h"
-#include "net/rpl/rpl-ns.h"
+#include "net/rpl-classic/rpl-private.h"
+#include "net/rpl-classic/rpl-ns.h"
 #include "lib/list.h"
 #include "lib/memb.h"
